Renamed plasma mp directory to match module name
