+++ micropython/modules/micropython.cmake
@@ -32,5 +32,5 @@ include(pico_unicorn/micropython)
 include(pico_display/micropython)
 include(pico_explorer/micropython)
 include(pico_wireless/micropython)
-include(plasma_2040/micropython)
+include(plasma/micropython)
 include(ulab/code/micropython)
