+++ micropython/modules/plasma/micropython.cmake
@@ -1,4 +1,4 @@
-set(MOD_NAME plasma_2040)
+set(MOD_NAME plasma)
 string(TOUPPER ${MOD_NAME} MOD_NAME_UPPER)
 add_library(usermod_${MOD_NAME} INTERFACE)
 
@@ -14,7 +14,6 @@ pico_generate_pio_header(usermod_${MOD_NAME} ${CMAKE_CURRENT_LIST_DIR}/../../../
 target_include_directories(usermod_${MOD_NAME} INTERFACE
     ${CMAKE_CURRENT_LIST_DIR}
     ${CMAKE_CURRENT_LIST_DIR}/../../../drivers/plasma/
-    ${CMAKE_CURRENT_LIST_DIR}/../../../libraries/plasma2040/
 )
 
 target_compile_definitions(usermod_${MOD_NAME} INTERFACE
